pyproject.toml
[tool.poetry]
name = "envirosentinel"
version = "0.1.0"
description = "IoT-based anomaly detection using environmental sensor data"
authors = ["Jonas Limpert <33454605+JonaLimp@users.noreply.github.com>"]
readme = "README.md"
packages = [
  { include = "envirosentinel" }
]

[tool.poetry.scripts]
envirosentinel = "run:app"

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
pandas = ">=2.2.3,<3.0.0"
scikit-learn = ">=1.6.1,<2.0.0"
requests = ">=2.32.3,<3.0.0"
flask = ">=3.1.0,<4.0.0"
python-dotenv = "^1.1.0"

[tool.poetry.group.dev.dependencies]
black = ">=25.1.0,<26.0.0"
flake8 = ">=7.2.0,<8.0.0"
mypy = ">=1.15.0,<2.0.0"
isort = ">=6.0.1,<7.0.0"
types-requests = "^2.32.0.20250328"
pyyaml = "^6.0.2"
ipykernel = "^6.29.5"
jupyter = "^1.1.1"
matplotlib = "^3.10.1"
seaborn = "^0.13.2"
missingno = "^0.5.2"
pytest = "^8.3.5"
pytest-flask = "^1.3.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
strict = false
mypy_path = ["data_ingest/src, predictor/src"]
explicit_package_bases = true
namespace_packages = true

[pytest]
python_paths = ["predictor/src", "data_ingest/src", "shared"]
